+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Kotlin 2.1.10
   - Bip39 1.0.9
   - Other dependencies update
+- Migrated to `zcash_client_backend 0.18.0`, `zcash_client_sqlite 0.16.0`
+- Added support for gap-limit-based discovery of transparent wallet addresses.
 
 ## [2.2.8] - 2025-03-03