+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this library adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for gap-limit-based discovery of transparent wallet addresses.
 - The internal `fetch-utxos` logic is now triggered only in every `init` and `complete` block sync phases, and it 
   fetches UTXOs from height 0 to support the Ledger funds rescue requirement.
+- Checkpoints update
 
 ## [2.2.8] - 2025-03-03